// apps/widget/src/interfaces/Settings.ts
export interface Settings {
    // Style Options
    backgroundColor: string;
    selectionColor: string;
    strokeColor: string;
    strokeWidth: number;
    cornerRadius: number;

    // Text Options
    labelText: string;

    // Visibility
    displayStatus: boolean;

    // Price Options
    firstPrice: number;
    secondPrice: number;

}